backend/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
django-celery-beat==2.9.0
django-guardian==2.4.0
django-fsm-2==4.2.4
psycopg[binary]==3.2.13
redis==5.3.1
celery==5.5.3
psutil==7.0.0
watchdog==5.0.3
